pyproject.toml
[tool.poetry]
name = "usage-overarch-halogen"
version = "0.1.0"
description = "Bitcoin Price Tracking Application"
authors = ["Victor R Zambrano"]
readme = "README.md"
packages = [{include = "*", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
# Web Framework
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.5.0"

# Database
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"

# Machine Learning
xgboost = "^2.0.3"
mlflow = "^2.14.1"
pandas = "^2.2.2"
numpy = "^2.0.0"
scikit-learn = "^1.5.0"

# Cloud Storage
boto3 = "^1.34.125"

# Utilities
requests = "^2.31.0"
python-dotenv = "^1.0.1"

# Dashboard
streamlit = "^1.31.0"
plotly = "^5.18.0"


[tool.poetry.scripts]
# API
start-api = "main:main"

# Model Training
train-price-model = "scripts.train_model:main"
train-trend-model = "scripts.train_trend_model:main"

# Maintenance
cleanup-predictions = "scripts.cleanup_old_predictions:main"

# Examples
examples = "scripts.example_usage:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
